+++ lib/src/build_tracker.dart
@@ -42,6 +42,8 @@ class BuildTracker {
   BuildTracker({
     this.onBuildFrame,
     this.printBuildFrame = true,
+    this.printBuildFrameIncludeBuilt = true,
+    this.printBuildFrameIncludeBuildScheduled = true,
     bool enabled = true,
   }) {
     this.enabled = enabled;
@@ -51,6 +53,16 @@ class BuildTracker {
   /// Print markdown-formatted stats after every frame.
   ///
   bool printBuildFrame;
+
+  ///
+  /// Print every widget that was build.
+  ///
+  bool printBuildFrameIncludeBuilt;
+
+  ///
+  /// Print every widget that was marked as dirty.
+  ///
+  bool printBuildFrameIncludeBuildScheduled;
 
   ///
   /// Called after every frame with `BuildFrame` information collected during the frame build.
@@ -65,8 +77,10 @@ class BuildTracker {
     }
     _enabled = value;
     if (value) {
-      assert(debugOnRebuildDirtyWidget == null, "`debugOnRebuildDirtyWidget` already in use ($debugOnRebuildDirtyWidget)");
-      assert(debugOnScheduleBuildFor == null, "`debugOnScheduleBuildFor` already in use ($debugOnScheduleBuildFor)");
+      assert(debugOnRebuildDirtyWidget == null,
+          "`debugOnRebuildDirtyWidget` already in use ($debugOnRebuildDirtyWidget)");
+      assert(debugOnScheduleBuildFor == null,
+          "`debugOnScheduleBuildFor` already in use ($debugOnScheduleBuildFor)");
       assert(
         WidgetsBinding.instance is TrackingBuildOwnerWidgetsBindingMixin,
         "`TrackingBuildOwnerWidgetsBindingMixin` is required (${WidgetsBinding.instance})",
@@ -81,35 +95,38 @@ class BuildTracker {
     }
   }
 
-  var _enabled = false;
-  var _number = 1;
-
-  var _frameCallbackScheduled = false;
-
-  final _buildList = <BuildEntry>[];
-  final _buildScheduleList = <ScheduleBuildEntry>[];
-
   void _onDebugOnRebuildDirtyWidget(Element e, bool builtOnce) {
-    _buildList.add(BuildEntry(timestamp: DateTime.now().millisecondsSinceEpoch, widget: e.debugGetCreatorChain(10)));
+    _buildList.add(BuildEntry(
+        timestamp: DateTime.now().millisecondsSinceEpoch,
+        widget: e.debugGetCreatorChain(10)));
   }
 
   void _onDebugOnScheduleBuildFor(Element e) {
     var chain = Chain.current();
-    final setStateIndex =
-        chain.traces.first.frames.lastIndexWhere((_) => {'State.setState', 'Element.markNeedsBuild', 'HookState.setState'}.contains(_.member));
+    final setStateIndex = chain.traces.first.frames.lastIndexWhere((_) => {
+          'State.setState',
+          'Element.markNeedsBuild',
+          'HookState.setState'
+        }.contains(_.member));
     if (setStateIndex > 0) {
       chain = Chain.current(setStateIndex);
     }
     _buildScheduleList.add(ScheduleBuildEntry(
         timestamp: DateTime.now().millisecondsSinceEpoch,
         widget: e.debugGetCreatorChain(10),
-        stack: chain.terse.traces.expand((t) => t.frames).map((f) => '${f.member},${f.location}').toBuiltList()));
+        stack: chain.terse.traces
+            .expand((t) => t.frames)
+            .map((f) => '${f.member},${f.location}')
+            .toBuiltList()));
   }
 
   void _frameCallback(Duration _) {
     _frameCallbackScheduled = false;
 
-    final frame = BuildFrame(number: _number++, built: BuiltList(_buildList), buildScheduled: BuiltList(_buildScheduleList));
+    final frame = BuildFrame(
+        number: _number++,
+        built: BuiltList(_buildList),
+        buildScheduled: BuiltList(_buildScheduleList));
     _buildList.clear();
     _buildScheduleList.clear();
 
@@ -132,7 +149,7 @@ class BuildTracker {
     debugPrint('# BuildTracker frame #${frame.number}');
     debugPrint('');
 
-    if (frame.built.isNotEmpty) {
+    if (printBuildFrameIncludeBuilt && frame.built.isNotEmpty) {
       debugPrint('## Widgets that were built');
       debugPrint('');
       for (final e in frame.built) {
@@ -141,23 +158,50 @@ class BuildTracker {
       debugPrint('');
     }
 
-    if (frame.buildScheduled.isNotEmpty) {
+    if (printBuildFrameIncludeBuildScheduled &&
+        frame.buildScheduled.isNotEmpty) {
       debugPrint('## Widgets that were marked dirty (build roots)');
       debugPrint('');
       for (final e in frame.buildScheduled) {
+        final stackHash = e.stack.hashCode.toRadixString(16);
+        final printedIn =
+            _printedStacks.putIfAbsent(e.stack, () => frame.number);
         debugPrint('### ${e.widget}:');
         debugPrint('');
-        debugPrint('```');
-        for (final frame in e.stack.map((_) => _.split(','))) {
-          final member = frame[0];
-          final location = frame[1];
-          final isCore = {'dart:', 'package:flutter/', 'package:flutter_test/'}.any((_) => location.startsWith(_));
-          debugPrint('${isCore ? ' ' : '*'} ${member.padRight(40)} $location');
+        if (printedIn != frame.number) {
+          debugPrint(
+              'Stack trace #$stackHash observed in frame #$printedIn for the first time');
+        } else {
+          debugPrint('Stack trace #$stackHash:');
+          debugPrint('```');
+          for (final frame in e.stack.map((_) => _.split(','))) {
+            final member = frame[0];
+            final location = frame[1];
+            final isCore = {
+              'dart:',
+              'package:flutter/',
+              'package:flutter_test/'
+            }.any((_) => location.startsWith(_));
+            debugPrint(
+                '${isCore ? ' ' : '*'} ${member.padRight(40)} $location');
+          }
+          debugPrint('```');
         }
-        debugPrint('```');
+        debugPrint('');
       }
     }
 
     debugPrint('# END of BuildTracker frame #${frame.number}');
+    debugPrint('');
   }
+
+  var _enabled = false;
+  var _number = 1;
+
+  var _frameCallbackScheduled = false;
+
+  final _buildList = <BuildEntry>[];
+  final _buildScheduleList = <ScheduleBuildEntry>[];
+
+  final _printedStacks = <BuiltList<String>, int>{};
 }