+++ lib/src/test_tracking_build_owner.dart
@@ -5,9 +5,13 @@ import 'package:flutter_test/flutter_test.dart';
 ///
 /// [AutomatedTestWidgetsFlutterBinding] with [TrackingBuildOwnerWidgetsBindingMixin]
 ///
-class TrackingBuildOwnerAutomatedTestWidgetsFlutterBinding extends AutomatedTestWidgetsFlutterBinding with TrackingBuildOwnerWidgetsBindingMixin {
+class TrackingBuildOwnerAutomatedTestWidgetsFlutterBinding
+    extends AutomatedTestWidgetsFlutterBinding
+    with TrackingBuildOwnerWidgetsBindingMixin {
   static WidgetsBinding ensureInitialized() {
-    if (WidgetsBinding.instance == null) TrackingBuildOwnerAutomatedTestWidgetsFlutterBinding();
+    if (WidgetsBinding.instance == null) {
+      TrackingBuildOwnerAutomatedTestWidgetsFlutterBinding();
+    }
     return WidgetsBinding.instance!;
   }
 }