+++ publish.sh
@@ -1,0 +1,3 @@
+./format.sh
+flutter analyze
+dart pub publish