+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.1
 
-First release with `BuildTracker`
+First release with `BuildTracker`.
 
 ## 0.0.2
 
@@ -8,4 +8,9 @@ Cleanup of dependencies to be null-safety compliant.
 
 ## 0.0.2+1
 
-Expose `BuildTracker`
+Expose `BuildTracker`.
+
+## 0.0.2+2
+
+- Add options to reduce output.
+- Show each stack trace only once.